[Adjust court/Muni relationship] Updated data model and logic to support municipalities having multiple courts; enabled sql file loading; changed citation search to use municipality ids vs names

+++ src/main/java/svc/models/Citation.java
@@ -3,8 +3,7 @@ package svc.models;
 import java.util.Date;
 import java.util.List;
 
-public class Citation 
-{
+public class Citation {
 	public int id;
 	public String citation_number;
 	public Date citation_date;
@@ -20,5 +19,5 @@ public class Citation
 	public String court_address;
 	
 	public List<Violation> violations;
-	public int court_id;
+	public Long court_id;
 }
